+++ include/utils/utils_extended.hpp
@@ -24,8 +24,8 @@ inline float compute_class_accuracy(const Tensor<float> &predictions,
   throw std::runtime_error("Unsupported device type for compute_class_accuracy.");
 }
 
-inline float compute_class_corrects(const Tensor<float> &predictions, const Tensor<float> &targets,
-                                    float threshold = 0.5f) {
+inline int compute_class_corrects(const Tensor<float> &predictions, const Tensor<float> &targets,
+                                  float threshold = 0.5f) {
   const size_t batch_size = predictions.shape()[0];
   const size_t num_classes = predictions.shape()[1];
